Site live on Google App Engine
From the React.JS build, the title of the directory was changed to EscatelTaxLaw so that the path of static files is preserved. The file app.yaml is read by google's app engine which specifies the static directory and app. A simple python script was used to host index.html. The website can be viewed live at escateltaxlaw.appspot.com.